Add utils crate
During the development of the contracts, we are starting to see patterns in code that gets repeated. This is a place were we can move that common logic (testing helper, conversions ...), without entangling the dependencies among contracts.

This can sound as a "disaster wardrobe" , and well, it is. But is a conscious one, that will be splitted in other modules with higher definition in terms of purpose once we have enough amount of logic and they deserve their own module.